remove cloning tests
